+++ src/main/kotlin/me/abacate/animefoda/request/NewCommentRequest.kt
@@ -3,7 +3,7 @@ package me.abacate.animefoda.request
 import java.util.*
 
 data class NewCommentRequest (
-    val parentId: Int?,
+    val parentId: Long?,
     val pageId: UUID,
     val content: String,
 )
